# lib/api-zod/src/generated/types/imageUploadInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImageUploadInput {
  filename: string;
  mimeType: string;
  /** @minimum 1 */
  sizeBytes: number;
}