# lib/api-zod/src/generated/types/imageUploadResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImageUploadResponse {
  uploadURL: string;
  objectPath: string;
}